+++ rotkehlchen/chain/ethereum/modules/yearn/constants.py
@@ -13,5 +13,6 @@ YEARN_LABEL_V3: Final = 'Yearn V3'
 YEARN_ICON: Final = 'yearn_vaults.svg'
 
 CPT_YEARN_STAKING: Final = 'yearn_staking'
+YEARN_STAKING_LABEL: Final = 'Yearn Staking'
 
 YEARN_PARTNER_TRACKER: Final = string_to_evm_address('0x8ee392a4787397126C163Cb9844d7c447da419D8')